refactor: remove old Connect to GitHub button and it into command bar
