+++ apps/nextjs/.env.local.sample
@@ -49,3 +49,4 @@ WIDGET_JWT_SECRET="set_me"
 FIRECRAWL_API_KEY="set_me"
 GITHUB_CLIENT_ID="set_me"
 GITHUB_CLIENT_SECRET="set_me"
+STRIPE_PRICE_ID="set_me"
