+++ apps/nextjs/src/app/(dashboard)/mailboxes/[mailbox_slug]/(inbox)/_components/messageActions.tsx
@@ -16,7 +16,6 @@ import { RouterOutputs } from "@/trpc";
 import { api } from "@/trpc/react";
 import { useUndoneEmailStore } from "./conversation";
 import { useConversationListContext } from "./conversationListContext";
-import { GitHubIssueButton } from "./githubIssueButton";
 import { useConversationsListInput } from "./shared/queries";
 import { TicketCommandBar } from "./ticketCommandBar";
 
@@ -267,14 +266,6 @@ export const MessageActions = () => {
                     ⌘K
                   </kbd>
                 </Button>
-                {conversation && (
-                  <GitHubIssueButton
-                    mailboxSlug={mailboxSlug}
-                    conversationSlug={conversation.slug}
-                    conversationSubject={conversation.subject || ""}
-                    conversationSummary={conversation.summary}
-                  />
-                )}
               </>
             ))}
         </div>
