+++ apps/nextjs/src/app/(dashboard)/mailboxes/[mailbox_slug]/(inbox)/_components/ticketCommandBar/mainPage.tsx
@@ -15,12 +15,13 @@ import { useConversationContext } from "@/app/(dashboard)/mailboxes/[mailbox_slu
 import useKeyboardShortcut from "@/components/useKeyboardShortcut";
 import { useToolExecution } from "@/hooks/useToolExecution";
 import { api } from "@/trpc/react";
+import GitHubSvg from "../../_components/icons/github.svg";
 import { CommandGroup } from "./types";
 
 type MainPageProps = {
   onGenerateDraft: () => void;
   onOpenChange: (open: boolean) => void;
-  setPage: (page: "main" | "previous-replies" | "assignees" | "notes" | "tools") => void;
+  setPage: (page: "main" | "previous-replies" | "assignees" | "notes" | "tools" | "github-issue") => void;
   setSelectedItemId: (id: string | null) => void;
   onToggleCc: () => void;
 };
@@ -40,6 +41,13 @@ export const useMainPage = ({
     { mailboxSlug, conversationSlug },
     { staleTime: Infinity, refetchOnMount: false, refetchOnWindowFocus: false, enabled: !!conversationSlug },
   );
+
+  const { data: mailbox } = api.mailbox.get.useQuery(
+    { mailboxSlug },
+    { staleTime: Infinity, refetchOnMount: false, refetchOnWindowFocus: false, enabled: !!mailboxSlug },
+  );
+
+  const isGitHubConnected = mailbox?.githubConnected && mailbox.githubRepoOwner && mailbox.githubRepoName;
 
   useKeyboardShortcut("n", (e) => {
     e.preventDefault();
@@ -143,6 +151,34 @@ export const useMainPage = ({
                 </p>
               </div>
             ),
+          },
+          {
+            id: "github-issue",
+            label: "GitHub Issue",
+            icon: GitHubSvg,
+            onSelect: () => {
+              setPage("github-issue");
+              setSelectedItemId(null);
+            },
+            shortcut: "G",
+            preview: (
+              <div className="p-4">
+                <h3 className="font-medium mb-2">GitHub Issue</h3>
+                {(conversation as any)?.githubIssueNumber ? (
+                  <>
+                    <p className="text-sm text-muted-foreground mb-2">
+                      This conversation is linked to GitHub issue #{(conversation as any).githubIssueNumber}.
+                    </p>
+                    <p className="text-sm text-muted-foreground">You can view the issue details, close or reopen it.</p>
+                  </>
+                ) : (
+                  <p className="text-sm text-muted-foreground">
+                    Create a new GitHub issue or link an existing one to this conversation.
+                  </p>
+                )}
+              </div>
+            ),
+            hidden: !isGitHubConnected,
           },
         ],
       },
@@ -251,6 +287,7 @@ export const useMainPage = ({
                   ),
                   onSelect: () => {
                     setPage("tools");
+                    setSelectedItemId(null);
                   },
                 },
               ],
@@ -258,7 +295,18 @@ export const useMainPage = ({
           ]
         : []),
     ],
-    [onGenerateDraft, onOpenChange, conversation, tools?.recommended, onToggleCc],
+    [
+      conversation?.status,
+      updateStatus,
+      onOpenChange,
+      setPage,
+      setSelectedItemId,
+      tools,
+      handleToolExecution,
+      onToggleCc,
+      onGenerateDraft,
+      isGitHubConnected,
+    ],
   );
 
   return mainCommandGroups;
